update integration test to use new polymorphic types
